added more works controller tests

+++ app/controllers/works_controller.rb
@@ -54,25 +54,47 @@ class WorksController < ApplicationController
 
   def show
     @work = find_work
+    if @work.nil?
+      head :not_found
+    end
   end
 
   def edit
     @work = find_work
+    if @work.nil?
+      head :not_found
+    end
   end
 
   def update
     @work = find_work
-    @work.update_attributes(works_params)
+    return head :not_found if @work.nil?
 
-    redirect_to work_path(@work.id)
+    if @work.update_attributes(works_params)
+      flash[:successful_edit] = "Updated #{@work.title}"
+      redirect_to work_path(@work.id)
+    else
+      flash.now[:failure_update] = "A problem occured: Could not update #{@work.category}"
+      if @work.category == "movie"
+        render :edit, status: :bad_request
+      elsif @work.category == "book"
+        render :edit, status: :bad_request
+      elsif @work.category == "album"
+        render :edit, status: :bad_request
+      end
+    end
   end
 
   def destroy
     work = find_work
-    path = find_path(work)
-    work.destroy
-    flash[:success] = "#{work.title} was deleted"
-    redirect_to path
+    if work.nil?
+      head :not_found
+    else
+      path = find_path(work)
+      work.destroy
+      flash[:success] = "#{work.title} was deleted"
+      redirect_to path
+    end
   end
 
 private
@@ -85,7 +107,7 @@ private
   end
 
   def find_work
-    Work.find(params[:id])
+    Work.find_by(id: params[:id])
   end
 
   def find_movies
